ft(delete-question-frontend): consumes the delete question endpoint - add js login - add delete button to question - fix delete endpoint [Finishes #160358895]

+++ src/controllers/question.js
@@ -111,33 +111,42 @@ class Question {
    */
   static destroy(request, response) {
     const id = request.params.questionId * 1;
+    const { userId } = request;
+
     db.query(`
     SELECT * 
     FROM questions 
     WHERE id = $1`, [id])
       .then((selectQueryResult) => {
-        if (selectQueryResult.rowCount !== 0) {
-          db.query('DELETE FROM questions where id = $1', [id])
-            .then((deleteQueryResult) => {
-              if (deleteQueryResult.rowCount === 1) {
-                return response.status(200).json({
-                  status: 'Success',
-                  message: 'Question deleted successfully'
+        if (userId === selectQueryResult.rows[0].user_id) {
+          if (selectQueryResult.rowCount !== 0) {
+            db.query('DELETE FROM questions where id = $1', [id])
+              .then((deleteQueryResult) => {
+                if (deleteQueryResult.rowCount === 1) {
+                  return response.status(200).json({
+                    status: 'Success',
+                    message: 'Question deleted successfully'
+                  });
+                }
+                return response.status(500).json({
+                  status: 'Failure',
+                  message: 'Something went wrong. Contact your administrator'
                 });
-              }
-              return response.status(500).json({
+              })
+              .catch(error => response.status(500).json({
                 status: 'Failure',
-                message: 'Something went wrong. Contact your administrator'
-              });
-            })
-            .catch(error => response.status(500).json({
+                message: 'Internal server error',
+              }));
+          } else {
+            return response.status(404).json({
               status: 'Failure',
-              message: 'Internal server error',
-            }));
+              message: 'Question not found'
+            });
+          }
         } else {
-          return response.status(404).json({
+          return response.status(403).json({
             status: 'Failure',
-            message: 'Question not found'
+            message: 'You are not authorized to modify this resource'
           });
         }
       })
